+++ llama_stack/providers/impls/meta_reference/evals/tasks/run_eval_task.py
@@ -3,7 +3,7 @@
 #
 # This source code is licensed under the terms described in the LICENSE file in
 # the root directory of this source tree.
-from llama_stack.distribution.registry.datasets.dataset_registry import DatasetRegistry
+from llama_stack.distribution.registry.datasets import DatasetRegistry
 from llama_stack.providers.impls.meta_reference.evals.scorer.basic_scorers import *  # noqa: F403
 from llama_stack.providers.impls.meta_reference.evals.generator.inference_generator import (
     InferenceGenerator,
@@ -38,9 +38,8 @@ class RunEvalTask(BaseTask):
     ) -> EvalResult:
         print(f"Running eval task w/ {eval_task_config}")
 
-        dataset = DatasetRegistry.get_dataset(
-            eval_task_config.dataset_config.dataset_name
-        )
+        print(DatasetRegistry.names())
+        dataset = DatasetRegistry.get(eval_task_config.dataset_config.dataset_name)
         dataset.load(n_samples=eval_task_config.dataset_config.row_limit)
         print(f"Running on {len(dataset)} samples")
